Move the config
